+++ src/frontends/onnx/frontend/src/core/graph.cpp
@@ -114,7 +114,7 @@ Graph::Graph(const std::shared_ptr<ONNX_NAMESPACE::ModelProto>& model_proto,
 
     if (m_extensions.telemetry) {
         for (const auto& op : op_statistics) {
-            m_extensions.telemetry->send_event("op_count", "onnx_" + op.first, op.second);
+            m_extensions.telemetry->send_event("op_count", "onnx_" + op.first, static_cast<int>(op.second));
         }
     }
 
@@ -148,7 +148,7 @@ void Graph::convert_to_ngraph_nodes() {
         }
         OutputVector ng_nodes{make_ng_nodes(node)};
         ++completed;
-        m_extensions.progress_reporter->report_progress(completed / total, total, completed);
+        m_extensions.progress_reporter->report_progress(completed / total, static_cast<unsigned int>(total), completed);
     }
 }
 
@@ -223,10 +223,10 @@ void Graph::decode_to_framework_nodes() {
         // Some of them may be optional and trimmed. See:
         // https://github.com/onnx/onnx/blob/master/docs/IR.md#optional-inputs-and-outputs
         for (std::size_t i{0}; i < node.get_outputs_size(); ++i) {
-            m_cache->emplace_node(node.output(i), std::move(ng_nodes.at(i)));
+            m_cache->emplace_node(node.output(static_cast<int>(i)), std::move(ng_nodes.at(i)));
         }
         ++completed;
-        m_extensions.progress_reporter->report_progress(completed / total, total, completed);
+        m_extensions.progress_reporter->report_progress(completed / total, static_cast<unsigned int>(total), completed);
     }
 }
 
@@ -236,7 +236,7 @@ std::shared_ptr<Function> Graph::create_function() {
     for (std::size_t i{0}; i < function->get_output_size(); ++i) {
         // the suffix makes the Result's name unique in case the nodes in the model don't have a name
         auto ov_result = function->get_output_op(i);
-        ov_result->set_friendly_name(detail::generate_result_name(onnx_outputs.Get(i).name(), ov_result));
+        ov_result->set_friendly_name(detail::generate_result_name(onnx_outputs.Get(static_cast<int>(i)).name(), ov_result));
     }
     return function;
 }
@@ -293,7 +293,7 @@ OutputVector Graph::make_ng_nodes(const Node& onnx_node) {
 
     for (std::size_t i{0}; i < onnx_node.get_outputs_size(); ++i) {
         auto ng_node_output = ng_subgraph_outputs.at(i);
-        m_cache->emplace_node(onnx_node.output(i), std::move(ng_node_output));
+        m_cache->emplace_node(onnx_node.output(static_cast<int>(i)), std::move(ng_node_output));
     }
 
     return ng_subgraph_outputs;
@@ -302,8 +302,8 @@ OutputVector Graph::make_ng_nodes(const Node& onnx_node) {
 void Graph::set_friendly_names(const Node& onnx_node, const OutputVector& ng_subgraph_outputs) const {
     if (onnx_node.op_type() == "Identity") {
         for (size_t i = 0; i < ng_subgraph_outputs.size(); ++i) {
-            ng_subgraph_outputs[i].get_tensor().add_names({onnx_node.output(i)});
-            ng_subgraph_outputs[i].get_node_shared_ptr()->set_friendly_name(onnx_node.output(i));
+            ng_subgraph_outputs[i].get_tensor().add_names({onnx_node.output(static_cast<int>(i))});
+            ng_subgraph_outputs[i].get_node_shared_ptr()->set_friendly_name(onnx_node.output(static_cast<int>(i)));
         }
         return;
     }
@@ -315,13 +315,13 @@ void Graph::set_friendly_names(const Node& onnx_node, const OutputVector& ng_sub
             break;
         }
 
-        ng_subgraph_outputs[i].get_node()->set_friendly_name(onnx_node.output(i));
+        ng_subgraph_outputs[i].get_node()->set_friendly_name(onnx_node.output(static_cast<int>(i)));
 
         // null node does not have tensor
         if (!ngraph::op::is_null(ng_subgraph_outputs[i])) {
-            ng_subgraph_outputs[i].get_tensor().set_names({onnx_node.output(i)});
+            ng_subgraph_outputs[i].get_tensor().set_names({onnx_node.output(static_cast<int>(i))});
             NGRAPH_SUPPRESS_DEPRECATED_START
-            ng_subgraph_outputs[i].get_tensor().set_name(onnx_node.output(i));
+            ng_subgraph_outputs[i].get_tensor().set_name(onnx_node.output(static_cast<int>(i)));
             NGRAPH_SUPPRESS_DEPRECATED_END
         }
     }
@@ -387,7 +387,7 @@ OutputVector Subgraph::make_framework_nodes(const Node& onnx_node) {
 
 void Subgraph::replace_input_from_parent_scope_with_parameter(const Node& onnx_node) {
     for (std::size_t i = 0; i < onnx_node.get_inputs_size(); ++i) {
-        const auto& in_name = onnx_node.input(i);
+        const auto& in_name = onnx_node.input(static_cast<int>(i));
         if (m_parent_graph->is_ng_node_in_cache(in_name) &&
             std::find(m_inputs_from_parent.begin(), m_inputs_from_parent.end(), in_name) ==
                 m_inputs_from_parent.end()) {
